# cloud_website/openapi.yaml
openapi: 3.0.3
info:
  title: Anywheredoor API
  description: |
    API for the Anywheredoor learning platform, providing authentication, course enrollment, 
    progress tracking, and administrative functions.
    
    ## Authentication
    
    - **User Endpoints**: Require NextAuth.js session cookie
    - **Admin Endpoints**: Require API key authentication with HMAC signature
    
    ## Rate Limiting
    
    - User endpoints: 100 requests/minute
    - Admin endpoints: 1000 requests/minute
  version: 1.0.0
  contact:
    email: api-support@anywheredoor.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://anywheredoor.com/api
    description: Production server
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for admin endpoints
    
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature for request verification
    
    TimestampAuth:
      type: apiKey
      in: header
      name: X-Timestamp
      description: Unix timestamp for request timing
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
      required:
        - error
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        priceCents:
          type: integer
        currency:
          type: string
        published:
          type: boolean
        featured:
          type: boolean
        level:
          type: string
        durationMin:
          type: integer
        rating:
          type: number
          format: float
        thumbnailUrl:
          type: string
          format: uri
        instructor:
          $ref: '#/components/schemas/Instructor'
        category:
          $ref: '#/components/schemas/Category'
    Instructor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
          format: uri
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
    Enrollment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        courseId:
          type: string
        enrolledAt:
          type: string
          format: date-time
        lastAccessedAt:
          type: string
          format: date-time
        completionPercentage:
          type: number
          format: float
        status:
          type: string
          enum: [ACTIVE, CANCELLED, COMPLETED, REFUNDED]
        source:
          type: string
        lastUpdatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
paths:
  /enrollments:
    post:
      tags:
        - Enrollments
      summary: Create course enrollment
      description: Enroll the authenticated user in a course
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                courseId:
                  type: string
                courseSlug:
                  type: string
              required:
                - courseId
                - courseSlug
      responses:
        '200':
          description: Enrollment created or payment required
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                      enrollment:
                        $ref: '#/components/schemas/Enrollment'
                  - type: object
                    properties:
                      requiresPayment:
                        type: boolean
                      purchaseId:
                        type: string
                      amount:
                        type: integer
                      currency:
                        type: string
        '400':
          description: Already enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      tags:
        - Enrollments
      summary: Get user enrollments
      description: Get all enrollments for the authenticated user
      security:
        - SessionCookie: []
      responses:
        '200':
          description: List of enrollments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Enrollment'
                        - type: object
                          properties:
                            course:
                              $ref: '#/components/schemas/Course'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /courses:
    get:
      tags:
        - Courses
      summary: List courses
      description: Get list of all published courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
...
  /admin/analytics/students:
    get:
      tags:
        - Admin - Analytics
      summary: Get student statistics
      description: Get student analytics and metrics
      security:
        - ApiKeyAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      responses:
        '200':
          description: Student statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      totalStudents:
                        type: integer
                      activeStudents:
                        type: integer
                      newStudentsThisMonth:
                        type: integer
                      averageCoursesPerStudent:
                        type: number
                      averageCompletionRate:
                        type: number
  /admin/audit-logs:
    get:
      tags:
        - Admin - Audit
      summary: Get audit logs
      description: Get audit logs with filtering
      security:
        - ApiKeyAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: action
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      logs:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            userId:
                              type: string
                            action:
                              type: string
                            resourceType:
                              type: string
                            resourceId:
                              type: string
                            details:
                              type: object
                            ipAddress:
                              type: string
                            userAgent:
                              type: string
                            createdAt:
                              type: string
                              format: date-time
                      pagination:
                        $ref: '#/components/schemas/Pagination'